requirements.txt
weasyprint
markdown
pymdown-extensions
pygments
pyyaml
click
jinja
